requirements.txt
streamlit
supabase
requests
beautifulsoup4
pandas
pulp
python-dotenv
langchain>=0.0.300
openai>=1.0.0
huggingface-hub